General changes to destructors
- I changed the contactList destructor, it was causing a crash in
eclipse whenever I quit.  I just implemented one out of the textbook and
it seems to be working fine.
- I changed the display function to use Garrett's printList function, it
should print 'contact cards' for each contact in the list.
- Something strange happened with the firstName and lastName variables
in the base class such that the constructor was causing an error.  I
commented out the constructor for firstName and lastName for now.
- I'm also noticing some of my changes keep getting erased
(specifically, the destructor for personalContact).  I'm always pulling
before I change, then committing when I'm done. Is that what you guys
are doing?